move tests (mode->modeactivator)
